+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `PeopleCard`: Add `disabled` prop.
+
 ## [11.2.2] - 2022-06-20
 
 Fix: